+++ StagingManagementSystem/src/main/java/com/revature/services/InterviewQuestionServiceImpl.java
@@ -7,6 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.revature.entities.InterviewQuestion;
+import com.revature.exceptions.SmsCustomException;
+import com.revature.exceptions.badrequests.NullReferenceException;
 import com.revature.repositories.InterviewQuestionRepo;
 
 @Service
@@ -22,22 +24,32 @@ public class InterviewQuestionServiceImpl implements InterviewQuestionService {
 
 	@Override
 	public void add(InterviewQuestion interviewQ) {
-		interviewQuestionRepo.saveAndFlush(interviewQ);
+		interviewQ = interviewQuestionRepo.saveAndFlush(interviewQ);
 	}
 
 	@Override
-	public void delete(InterviewQuestion interviewQ) {
+	public void delete(InterviewQuestion interviewQ) throws SmsCustomException{
+		if(interviewQ == null){
+			throw new NullReferenceException("Interview Questions is null.");
+		}
 		interviewQuestionRepo.delete(interviewQ);
 	}
 
 	@Override
-	public void update(InterviewQuestion interviewQ) {
-		interviewQuestionRepo.saveAndFlush(interviewQ);
+	public void update(InterviewQuestion interviewQ) throws SmsCustomException{
+		if(interviewQ == null){
+			throw new NullReferenceException("Interview Question is null.");
+		}
+		interviewQ = interviewQuestionRepo.saveAndFlush(interviewQ);
 	}
 
 	@Override
-	public InterviewQuestion findById(long id) {
-		return interviewQuestionRepo.getOne(id);
+	public InterviewQuestion findById(long id) throws SmsCustomException{
+		InterviewQuestion interviewQ = interviewQuestionRepo.getOne(id);
+		if(interviewQ == null){
+			throw new NullReferenceException("Interview Question is null.");
+		}
+		return interviewQ;
 	}
 
 	@Override
